app/requirements.txt
fastapi
uvicorn
PyGithub
openai 